# Proyecto_personal(grafana_prometheus)/backend/rest/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Conduit API",
      "version": "1.0.0",
      "title": "Conduit API",
      "contact": {
        "name": "RealWorld",
        "url": "https://realworld.io"
      },
      "license": {
        "name": "MIT License",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": [
      "http"
    ],
    "produces": [
      "application/json"
    ],
    "consumes": [
      "application/json"
    ],
    "securityDefinitions": {
      "Token": {
        "description": "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token: xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "paths": {
...
      "/articles/feed": {
        "get": {
          "summary": "Get recent articles from users you follow",
          "description": "Get most recent articles from users you follow. Use query parameters to limit. Auth is required",
          "tags": [
            "Articles"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "GetArticlesFeed",
          "parameters": [
            {
              "name": "limit",
              "in": "query",
              "description": "Limit number of articles returned (default is 20)",
              "required": false,
              "default": 20,
              "type": "integer"
            },
            {
              "name": "offset",
              "in": "query",
              "description": "Offset/skip number of articles (default is 0)",
              "required": false,
              "default": 0,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/MultipleArticlesResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/articles": {
        "get": {
          "summary": "Get recent articles globally",
          "description": "Get most recent articles globally. Use query parameters to filter results. Auth is optional",
          "tags": [
            "Articles"
          ],
          "operationId": "GetArticles",
          "parameters": [
            {
              "name": "tag",
              "in": "query",
              "description": "Filter by tag",
              "required": false,
              "type": "string"
            },
            {
              "name": "author",
              "in": "query",
              "description": "Filter by author (username)",
              "required": false,
              "type": "string"
            },
            {
              "name": "favorited",
              "in": "query",
              "description": "Filter by favorites of a user (username)",
              "required": false,
              "type": "string"
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit number of articles returned (default is 20)",
              "required": false,
              "default": 20,
              "type": "integer"
            },
            {
              "name": "offset",
              "in": "query",
              "description": "Offset/skip number of articles (default is 0)",
              "required": false,
              "default": 0,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/MultipleArticlesResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "post": {
          "summary": "Create an article",
          "description": "Create an article. Auth is required",
          "tags": [
            "Articles"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "CreateArticle",
          "parameters": [
            {
              "name": "article",
              "in": "body",
              "required": true,
              "description": "Article to create",
              "schema": {
                "$ref": "#/definitions/NewArticleRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleArticleResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/articles/{slug}": {
        "get": {
          "summary": "Get an article",
          "description": "Get an article. Auth not required",
          "tags": [
            "Articles"
          ],
          "operationId": "GetArticle",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the article to get",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleArticleResponse"
              }
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "put": {
          "summary": "Update an article",
          "description": "Update an article. Auth is required",
          "tags": [
            "Articles"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "UpdateArticle",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the article to update",
              "type": "string"
            },
            {
              "name": "article",
              "in": "body",
              "required": true,
              "description": "Article to update",
              "schema": {
                "$ref": "#/definitions/UpdateArticleRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleArticleResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete an article",
          "description": "Delete an article. Auth is required",
          "tags": [
            "Articles"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "DeleteArticle",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the article to delete",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/articles/{slug}/comments": {
        "get": {
          "summary": "Get comments for an article",
          "description": "Get the comments for an article. Auth is optional",
          "tags": [
            "Comments"
          ],
          "operationId": "GetArticleComments",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the article that you want to get comments for",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/MultipleCommentsResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "post": {
          "summary": "Create a comment for an article",
          "description": "Create a comment for an article. Auth is required",
          "tags": [
            "Comments"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "CreateArticleComment",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the article that you want to create a comment for",
              "type": "string"
            },
            {
              "name": "comment",
              "in": "body",
              "required": true,
              "description": "Comment you want to create",
              "schema": {
                "$ref": "#/definitions/NewCommentRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleCommentResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/articles/{slug}/comments/{id}": {
        "delete": {
          "summary": "Delete a comment for an article",
          "description": "Delete a comment for an article. Auth is required",
          "tags": [
            "Comments"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "DeleteArticleComment",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the article that you want to delete a comment for",
              "type": "string"
            },
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the comment you want to delete",
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/articles/{slug}/favorite": {
        "post": {
          "summary": "Favorite an article",
          "description": "Favorite an article. Auth is required",
          "tags": [
            "Favorites"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "CreateArticleFavorite",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the article that you want to favorite",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleArticleResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "delete": {
          "summary": "Unfavorite an article",
          "description": "Unfavorite an article. Auth is required",
          "tags": [
            "Favorites"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "DeleteArticleFavorite",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the article that you want to unfavorite",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleArticleResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Profile": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "following": {
            "type": "boolean"
          }
        },
        "required": [
          "username",
          "bio",
          "image",
          "following"
        ]
      },
      "Article": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "tagList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "favorited": {
            "type": "boolean"
          },
          "favoritesCount": {
            "type": "integer"
          },
          "author": {
            "$ref": "#/definitions/Profile"
          }
        },
        "required": [
          "slug",
          "title",
          "description",
          "body",
          "tagList",
          "createdAt",
          "updatedAt",
          "favorited",
          "favoritesCount",
          "author"
        ]
      },
      "SingleArticleResponse": {
        "type": "object",
        "properties": {
          "article": {
            "$ref": "#/definitions/Article"
          }
        },
        "required": [
          "article"
        ]
      },
      "MultipleArticlesResponse": {
        "type": "object",
        "properties": {
          "articles": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Article"
            }
          },
          "articlesCount": {
            "type": "integer"
          }
        },
        "required": [
          "articles",
          "articlesCount"
        ]
      },
      "NewArticle": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "tagList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "description",
          "body"
        ]
      },
      "NewArticleRequest": {
        "type": "object",
        "properties": {
          "article": {
            "$ref": "#/definitions/NewArticle"
          }
        },
        "required": [
          "article"
        ]
      },
      "UpdateArticle": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
      "UpdateArticleRequest": {
        "type": "object",
        "properties": {
          "article": {
            "$ref": "#/definitions/UpdateArticle"
          }
        },
        "required": [
          "article"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "body": {
            "type": "string"
          },
          "author": {
            "$ref": "#/definitions/Profile"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "body",
          "author"
        ]
      },
      "SingleCommentResponse": {
        "type": "object",
        "properties": {
          "comment": {
            "$ref": "#/definitions/Comment"
          }
        },
        "required": [
          "comment"
        ]
      },
      "MultipleCommentsResponse": {
        "type": "object",
        "properties": {
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Comment"
            }
          }
        },
        "required": [
          "comments"
        ]
      },
      "NewComment": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        },
        "required": [
          "body"
        ]
      },
      "NewCommentRequest": {
        "type": "object",
        "properties": {
          "comment": {
            "$ref": "#/definitions/NewComment"
          }
        },
        "required": [
          "comment"
        ]
      },
...
      "GenericErrorModel": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "body": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "body"
            ]
          }
        },
        "required": [
          "errors"
        ]
      }
    }
  }